always bundle javascript library
means we can include the kernel/sections in the bundled javascript,
because we no longer need to support deno in dev.

also switch to tsc for typechecking, allowing declarations emit.

as a consequence, I've had to drop support for nodejs, because its
non-compliant Worker implementation, and its lack of fetch("file://...")
support makes it harder to support.