ci: add horrible hack to retry compilation on failure
it seems there's some race condition in the build system,
and the gen_init_cpio step fails on occasion.

because make is incremental, when the build is restarted it'll pick
straight back up at the failed step, and the problem will resolve.

a hacky fix sure, but good enough for now.